Allow leaving an MLS conversation via Wire API
