+++ services/galley/galley.cabal
@@ -34,6 +34,8 @@ library
     Galley.API.MLS.KeyPackage
     Galley.API.MLS.Keys
     Galley.API.MLS.Message
+    Galley.API.MLS.Propagate
+    Galley.API.MLS.Removal
     Galley.API.MLS.Welcome
     Galley.API.One2One
     Galley.API.Public
