+++ services/galley/src/Galley/API/Teams/Features.hs
@@ -52,10 +52,12 @@ import Galley.API.LegalHold (isLegalHoldEnabledForTeam)
 import qualified Galley.API.LegalHold as LegalHold
 import Galley.API.Teams (ensureNotTooLargeToActivateLegalHold)
 import Galley.API.Util (assertTeamExists, getTeamMembersForFanout, membersToRecipients, permissionCheck)
+import Galley.App
 import Galley.Effects
 import Galley.Effects.BrigAccess (getAccountConferenceCallingConfigClient, updateSearchVisibilityInbound)
 import Galley.Effects.ConversationStore as ConversationStore
 import Galley.Effects.GundeckAccess
+import Galley.Effects.ProposalStore
 import qualified Galley.Effects.SearchVisibilityStore as SearchVisibilityData
 import Galley.Effects.TeamFeatureStore
 import qualified Galley.Effects.TeamFeatureStore as TeamFeatures
@@ -726,10 +728,12 @@ instance SetFeatureConfig db LegalholdConfig where
              FireAndForget,
              GundeckAccess,
              Input (Local ()),
+             Input Env,
              Input UTCTime,
              LegalHoldStore,
              ListItems LegacyPaging ConvId,
              MemberStore,
+             ProposalStore,
              TeamFeatureStore db,
              TeamStore,
              TeamMemberStore InternalPaging,
